Some changes on blade

+++ database/migrations/2021_05_19_071513_create_student_rating_forms_table.php
@@ -52,7 +52,7 @@ class CreateStudentRatingFormsTable extends Migration
             $table->integer('management_of_learning_total');
             $table->longText('comments')->nullable();
             $table->integer('total');
-            $table->integer('scale');
+            $table->float('scale');
             $table->timestamps();
         });
     }
